refactor: rename task-updater to task-mutations
The module has exported the taskMutations namespace (create + update)
since vault_create_task landed; the filename still said updater.
