+++ src/vault-mcp/vault-operations/task-mutations.ts
